INNOVA/backend/requirements.txt
fastapi==0.115.0
gunicorn==23.0.0
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-multipart==0.0.9
sse-starlette==1.6.1
cohere==5.5.8
huggingface_hub>=0.35.0
transformers>=4.40.0
torch>=2.0.0
safetensors>=0.4.0
accelerate>=0.25.0
protobuf>=4.24
llama-cpp-python==0.2.90
peft>=0.12.0
